HvD: The icc compiler complained about using the %d format specifier in print statements which are being handed a long variable. So I have inserted the appropriate type casts to cast all offending variables to int.

+++ src/peigs/src/c/clustrf5.c
@@ -445,7 +445,8 @@ Integer clustrf5_ (n, d, e, m, w, mapZ, vecZ, iblock, nsplit, isplit, ptbeval, n
 	  eps1 = fabs(eps * xj);
 	  pertol = eps1 * R_TEN;
 	  if ((xj - w[j-1]) < -eps1) {
-	    printf(" Error in ordering eigenvalues: -5 error clustrf me = %d \n", me );
+	    printf(" Error in ordering eigenvalues: -5 error clustrf me = %d \n",
+                   (int) me );
 	  }
 	  sep = xj - xjm;
 /*
